feat(alarm-card): alarm-type badge nested in the corner, spinning as the card collapses
A zero-elevation round badge sits in the card's top-left corner showing the alarm's kind —
SemiCircle (early, <6), Sunny (morning, 6–9), Boom (late, ≥9). Its radius is concentric with
the card corner (Radius.xl) AND the collapsed pill's left circle (same radius), so it nests
pixel-perfectly in both states; it stays put and only spins like a cog as the card collapses,
while the clock shifts right to clear it. Extracted SleepTimeline to its own file to stay
under the line cap; drawMorph made internal so the badge reuses the fit/rotate draw.
